@essweine 2024-07-03
There was originally a reason for this (I think it had to do with error handling) but I don't think it's relevant at this point.